# users_api/OpenAPI.yaml
openapi: 3.0.0
info:
  title: Flask User API
  description: API for user management
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development server
paths:
...
  /users:
    get:
      summary: Get a list of users
      parameters:
        - name: page
          in: query
          description: "Page number for pagination (default: 1)"
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 1000
            format: int32
        - name: per_page
          in: query
          description: "Number of items per page (default: 10)"
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 1000
            format: int32
        - name: search_name
          in: query
          description: "Search filter by name (default: '')"
          required: false
          schema:
            type: string
            maxLength: 100
            pattern: ^[A-Za-z0-9]{3,10}$
      responses:
        '200':
          description: List of users
          content:
            application/json:
              example:
                page: 1
                per_page: 10
                search_name: John
                users:
                  schema:
                    - $ref: '#/components/schemas/Full-User'
              schema:
                $ref: '#/components/schemas/Paginado'
        '401':
          description: Credenciales invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '404':
          description: List of users
          content:
            application/json:
              example:
                page: 1
                per_page: 10
                search_name: Patricia
                users:
              schema:
                $ref: '#/components/schemas/Paginado'
        '406':
          description: Not Acceptable (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '412':
          description: El campo nombre es requerido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '415':
          description: Unsupported Media Type (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '429':
          description: Too Many Requests (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '500':
          description: Error al registrar usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
    
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Full-User'
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '401':
          description: Credenciales invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '406':
          description: Not Acceptable (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '412':
          description: El campo nombre es requerido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '415':
          description: Unsupported Media Type (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '429':
          description: Too Many Requests (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '500':
          description: Error al registrar usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP' 
  /users/{user_id}:
    get:
      summary: Get user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user to retrieve
          schema:
            type: integer
            minimum: 0
            maximum: 1000
            format: int32
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Full-User'
        '401':
          description: Credenciales invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '406':
          description: Not Acceptable (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '415':
          description: Unsupported Media Type (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '429':
          description: Too Many Requests (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '500':
          description: Error al obtener el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
    put:
      summary: Update logged in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Full-User'
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '500':
          description: Error al actualizar el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '401':
          description: Credenciales invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '406':
          description: Not Acceptable (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '415':
          description: Unsupported Media Type (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '429':
          description: Too Many Requests (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
    delete:
      summary: Delete logged in user
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '401':
          description: Token expirado o invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '406':
          description: Not Acceptable (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '415':
          description: Unsupported Media Type (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '429':
          description: Too Many Requests (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP' 
        '500':
          description: Error al eliminar el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP' 
  /users/{user_id}/new-password:
    patch:
      summary: Change user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nueva_clave:
                  type: string
                  maxLength: 100
                  pattern: ^[A-Za-z0-9]{3,10}$
              additionalProperties: false
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '401':
          description: Token expirado o invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '406':
          description: Not Acceptable (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '415':
          description: Unsupported Media Type (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '429':
          description: Too Many Requests (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '500':
          description: Error al actualizar la clave
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
  /auth:
    post:
      summary: User authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  maxLength: 100
                  pattern: ^[A-Za-z0-9]{3,10}$
                clave:
                  type: string
                  maxLength: 100
                  pattern: ^[A-Za-z0-9]{3,10}$
              additionalProperties: false
      responses:
        '200':
          description: JWT token generated successfully
          content:
            application/json:
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Credenciales invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '406':
          description: Not Acceptable (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '415':
          description: Unsupported Media Type (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '429':
          description: Too Many Requests (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '500':
          description: Error en el proceso de login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
  /auth/password-reset:
    post:
      summary: Forgot user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  maxLength: 50
                  pattern: ^[A-Za-z0-9]{3,10}$
              additionalProperties: false
      responses:
        '200':
          description: Password reset token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '404':
          description: Correo electrónico no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '406':
          description: Not Acceptable (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '415':
          description: Unsupported Media Type (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '429':
          description: Too Many Requests (en-US)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        '504':
          description: Error en el proceso de recuperación de clave
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodigoHTTP'
security:
  - bearerAuth: []
components:
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
  schemas:
    Full-User:
      $ref: './schemas/Full-User.json'
...
    CodigoHTTP:
      $ref: './schemas/Codigo-HTTP.json'
    Paginado:
      $ref: './schemas/Paginado.json'
    Token:
      $ref: './schemas/Token.json'